+++ actions.lua
@@ -1,24 +1,25 @@
 -- Actions
 -- Created By Jackz
 local SCRIPT = "actions"
-local VERSION = "1.7.3"
+local VERSION = "1.7.5"
 local CHANGELOG_PATH = filesystem.stand_dir() .. "/Cache/changelog_" .. SCRIPT .. ".txt"
 -- Check for updates & auto-update: 
 -- Remove these lines if you want to disable update-checks & auto-updates: (7-54)
-util.async_http_get("jackz.me", "/stand/updatecheck.php?ucv=2&script=" .. SCRIPT .. "&v=" .. VERSION, function(result)
+async_http.init("jackz.me", "/stand/updatecheck.php?ucv=2&script=" .. SCRIPT .. "&v=" .. VERSION, function(result)
     chunks = {}
     for substring in string.gmatch(result, "%S+") do
         table.insert(chunks, substring)
     end
     if chunks[1] == "OUTDATED" then
         -- Remove this block (lines 15-31) to disable auto updates
-        util.async_http_get("jackz.me", "/stand/changelog.php?raw=1&script=" .. SCRIPT .. "&since=" .. VERSION, function(result)
+        async_http.init("jackz.me", "/stand/changelog.php?raw=1&script=" .. SCRIPT .. "&since=" .. VERSION, function(result)
             local file = io.open(CHANGELOG_PATH, "w")
             io.output(file)
             io.write(result:gsub("\r", "") .. "\n") -- have to strip out \r for some reason, or it makes two lines. ty windows
             io.close(file)
         end)
-        util.async_http_get("jackz.me", "/stand/lua/" .. SCRIPT .. ".lua", function(result)
+        async_http.dispatch()
+        async_http.init("jackz.me", "/stand/lua/" .. SCRIPT .. ".lua", function(result)
             local file = io.open(filesystem.scripts_dir() .. "/" .. SCRIPT .. ".lua", "w")
             io.output(file)
             io.write(result:gsub("\r", "") .. "\n") -- have to strip out \r for some reason, or it makes two lines. ty windows
@@ -29,47 +30,37 @@ util.async_http_get("jackz.me", "/stand/updatecheck.php?ucv=2&script=" .. SCRIPT
             util.toast(SCRIPT .. ": Failed to automatically update to V" .. chunks[2] .. ".\nPlease download latest update manually.\nhttps://jackz.me/stand/get-latest-zip", 2)
             util.stop_script()
         end)
+        async_http.dispatch()
     end
 end)
--- Start Library Requirements
-local status, json = pcall(require, "json")
-if not status then
-    WaitingLibsDownload = true
-    util.async_http_get("jackz.me", "/stand/libs/json.lua", function(result)
-        local file = io.open(filesystem.scripts_dir() .. "/lib/json.lua", "w")
-        io.output(file)
-        io.write(result)
-        io.close(file)
-        WaitingLibsDownload = false
-        util.toast(SCRIPT .. ": Automatically downloaded missing lib 'json'")
-        json = require(file)
-    end, function(e)
-        util.toast(SCRIPT .. " cannot load: Library files are missing. (json)", 10)
-        util.stop_script()
-    end)
-end
+async_http.dispatch()
 local WaitingLibsDownload = false
-function try_load_lib(lib)
-    local status = pcall(require, lib)
+function try_load_lib(lib, globalName)
+    local status, f = pcall(require, string.sub(lib, 0, #lib - 4))
     if not status then
         WaitingLibsDownload = true
-        util.async_http_get("jackz.me", "/stand/libs/" .. lib .. ".lua", function(result)
-            local file = io.open(filesystem.scripts_dir() .. "/lib/" .. lib .. ".lua", "w")
+        async_http.init("jackz.me", "/stand/libs/" .. lib, function(result)
+            local file = io.open(filesystem.scripts_dir() .. "/lib/" .. lib, "w")
             io.output(file)
             io.write(result)
             io.close(file)
             WaitingLibsDownload = false
-            util.toast(SCRIPT .. ": Automatically downloaded missing lib '" .. lib .. ".lua'")
-            require(lib)
+            util.toast(SCRIPT .. ": Automatically downloaded missing lib '" .. lib .. "'")
+            if globalName then
+                _G[globalName] = require(string.sub(lib, 0, #lib - 4))
+            end
         end, function(e)
             util.toast(SCRIPT .. " cannot load: Library files are missing. (" .. lib .. ")", 10)
             util.stop_script()
         end)
+        async_http.dispatch()
+    elseif globalName then
+        _G[globalName] = f
     end
 end
-try_load_lib("natives-1627063482")
-try_load_lib("animations")
-
+try_load_lib("natives-1627063482.lua")
+try_load_lib("animations.lua")
+-- If script is actively downloading new update, wait:
 while WaitingLibsDownload do
     util.yield()
 end
@@ -353,6 +344,7 @@ local menus = {
     headers = {},
     subheaders = {}
 }
+menu.divider(animationsMenu, "Animations")
 for _, header in ipairs(ANIMATIONS_HEADINGS) do
     if not menus[header] then
         menus.headers[header] = menu.list(animationsMenu, header)